+++ app/routes/api/gallery.js
@@ -1,11 +1,11 @@
 const express = require("express"),
     router = express.Router(),
+    _ = require("lodash"),
     fsPromises = require("fs").promises,
     rename = fsPromises.rename,
     path = require("path"),
     multer = require("multer"),
     logger = require("./../../lib/logger"),
-    config = require("./../../lib/config"),
     store = require("./../../store"),
     uploadPath = path.resolve(process.cwd(), process.env.GALLERY_UPLOAD_PATH),
     rootPath = path.resolve(process.cwd(), process.env.GALLERY_ROOT_PATH),
@@ -15,46 +15,76 @@ const express = require("express"),
     upload = multer({ dest: uploadPath });
 
 router.get("/photosets", async function (req, res) {
-    let photoSets = await store.photoSets.all();
+    let photoSets = await store.photoSets.all({code: 1});
     res.json(photoSets);
 });
 
 router.get("/photos", async function (req, res) {
-    let photos = await store.photos.all();
+    let photos = await store.photos.all({created: 1});
     return res.json(photos);
 });
 
 router.post("/photo", upload.single("file"), async (req, res) => {
     let photo = req.body,
-        isNew = !photo._id;
+        isNew = !photo._id,
+        dbPhoto;
 
-    // parse photo _id and sets array
+    // parse photo _id
     photo._id = Number(photo._id);
     photo.sets = JSON.parse(photo.sets);
 
     // TODO: use mongoDB date
     if (isNew) {
+
         photo.created = new Date();
+
+        // if it's a new photo then save it to get _id
+        await store.photos.save(photo);
+
+        // update all related photoSets to include the new photo
+        await store.photoSets.updateMany(
+            { code: { $in: photo.sets } },
+            { $push: { photos: photo._id } }
+        );
+
     } else {
+
         photo.updated = new Date();
+
+        // get original photo from database
+        dbPhoto = await store.photos.getById(photo._id);
+
+        // update all related photoSets
+        let addToSets = _.difference(photo.sets,dbPhoto.sets),
+            removeFromSets = _.difference(dbPhoto.sets,photo.sets);
+
+        // add photo to `addToSets`
+        await store.photoSets.updateMany(
+            { code: { $in: addToSets } },
+            { $push: { photos: photo._id } }
+        );
+
+        // remove photo from `removeFromSets`
+        await store.photoSets.updateMany(
+            { code: { $in: removeFromSets } },
+            { $pull: { photos: photo._id } }
+        );
+
     }
 
+    // save file if present
     if (req.file) {
         const relativePhotosPath = path.relative(req.app.get("static-path"), photosPath),
             extension = path.extname(req.file.originalname).toLowerCase();
 
-        if (isNew) {
-            // if it's a new photo then save it to get _id
-            await store.photos.save(photo);
-        } else {
-            // if it's not a new photo then backup old photo file
-            let oldPhoto, oldPhotoPath;
+        if (!isNew) {
+            // if it's not a new photo then try to backup old photo file
+            let dbPhotoPath;
             try {
-                oldPhoto = await store.photos.getById(photo._id);
-                oldPhotoPath = path.resolve(req.app.get("static-path"), oldPhoto.src.slice(1));
-                await rename(oldPhotoPath, `${trashPath}/${path.basename(oldPhotoPath)}`);
+                dbPhotoPath = path.resolve(req.app.get("static-path"), dbPhoto.src.slice(1));
+                await rename(dbPhotoPath, `${trashPath}/${path.basename(dbPhotoPath)}`);
             } catch (err) {
-                logger.warn(`Can't backup photo file: ${oldPhotoPath}`);
+                logger.warn(`Can't backup photo file: ${dbPhotoPath}`);
             }
         }
 
@@ -64,7 +94,6 @@ router.post("/photo", upload.single("file"), async (req, res) => {
     }
 
     await store.photos.save(photo);
-    console.log(photo);
     return res.json(photo);
 });
 
@@ -92,6 +121,9 @@ router.post("/photoset", upload.single("file"), async (req, res) => {
         photoSet.cover = `/${relativeCoversPath}/${photoSet._id}${extension}`;
         await rename(req.file.path, `${coversPath}/${photoSet._id}${extension}`);
     }
+    if (isNew) {
+        photoSet.photos = [];
+    }
     await store.photoSets.save(photoSet, isNew);
     return res.json(photoSet);
 });
@@ -108,8 +140,16 @@ router.delete("/photo/:id", async (req, res) => {
     let photo = await store.photos.getById(_id);
     // if photo is in trash
     if (photo.deleted) {
+
+        // remove photo from related photosets
+        await store.photoSets.findOneAndUpdate(
+            { code: { $in: photo.sets } },
+            { $pull: { photos: _id } }
+        );
+
         // delete permanently
         await store.photos.delete(_id);
+
         // move related file to trash
         let src = path.resolve(req.app.get("static-path"), photo.src.slice(1));
         await rename(src, `${trashPath}/${path.basename(src)}`);
@@ -126,8 +166,28 @@ router.delete("/photo/:id", async (req, res) => {
 
 router.delete("/photoset/:id", async (req, res) => {
     let _id = req.params.id;
+    let photoSet = await store.photoSets.getById(_id);
+
+    // unlink related photos
+    await store.photos.updateMany(
+        { _id: { $in: photoSet.photos } },
+        { $pull: { sets: photoSet.code } }
+    );
+
+    // delete photoset permanently
     await store.photoSets.delete(_id);
     return res.json({ _id });
+});
+
+router.post("/photos/reorder", async (req, res) => {
+    console.log(req.body);
+    let photoSetCode = req.body.photoSet,
+        photos = req.body.photos;
+    await store.photoSets.findOneAndUpdate(
+        { code: photoSetCode },
+        { $set: { photos } }
+    );
+    return res.sendStatus(200);
 });
 
 module.exports = router;